created separate folder for styles
