+++ client/src/components/styles/Stars.jsx
